@@ -1,11 +1,5 @@
 source 'https://rubygems.org'
-ruby '2.7.5'
-
-# ref: https://github.com/rails/rails/commit/12d5c21031446686898d5bac924ff3e9e34b6a7d
-git_source(:github) do |repo_name|
-  repo_name = "#{repo_name}/#{repo_name}" unless repo_name.include?("/")
-  "https://github.com/#{repo_name}.git"
-end
+ruby '3.1.0'
 
 gem 'rails'
 gem 'bootstrap-sass'
@@ -17,10 +11,6 @@ gem 'jquery-rails'
 gem 'whenever', require: false
 gem 'haml'
 gem 'faraday'
-
-# Rails5.0.0.beta1 support is none: https://discuss.newrelic.com/t/rails-5-0-support/26595/10
-# gem 'newrelic_rpm'
-
 gem 'active_model_serializers'
 gem 'sprockets'
 gem 'sprockets-rails'
@@ -29,7 +19,6 @@ gem 'responders'
 gem 'coffee-rails'
 gem 'compass-rails', github: 'Compass/compass-rails'
 gem 'execjs'
-gem 'therubyracer'
 gem 'uglifier'
 gem 'rack-google-analytics'
 gem 'awesome_print'
@@ -48,7 +37,7 @@ group :development, :test do
   %w[rspec-core rspec-expectations rspec-mocks rspec-support].each do |lib|
     gem lib, :git => "https://github.com/rspec/#{lib}.git", :branch => 'master'
   end
-  gem 'rspec-rails', :git => "https://github.com/rspec/rspec-rails.git", :branch => '4-0-dev'
+  gem 'rspec-rails'
   gem 'annotate'
   gem 'factory_bot_rails'
   gem 'tapp'
